Move glyphs and arcs on canvas in Vanted...
Problem: After import of the SBGN-ML into Vanted, no error message is shown, but the map is not visible either.
Solution: The map is located out of the visible canvas in Vanted. All parts of the map are marked and moved to the visible area. Then SBGN-ML is exported again.